Переделал удаление и создал редактирование

+++ Source/webapp/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, get_object_or_404
 
 from webapp.models import ToDoList
 
@@ -17,18 +17,20 @@ def todolist_detail(request, *args, **kwargs):
     return render(request, "detail.html", {"todolist": todolist})
 
 
-def todolist_delete(request):
-    todo_id = request.GET.get('id')
-    todolist = ToDoList.objects.get(id=todo_id)
-    todolist.delete()
-    redirect('todolist', id=todolist.pk)
+def todolist_delete(request, pk):
+    todolist = get_object_or_404(ToDoList, id=pk)
+    if request.method == "GET":
+        return render(request, "delete.html")
+    else:
+        todolist.delete()
+        return redirect('todolist')
 
 
 def todolist_create(request):
     if request.method == "GET":
         return render(request, "create.html")
     if request.method == 'POST':
-        todolist = ToDoList.objects.create(
+        ToDoList.objects.create(
             description=request.POST.get('description'),
             detailed_description=request.POST.get('detailed_description'),
             status=request.POST.get('status'),
@@ -36,4 +38,19 @@ def todolist_create(request):
             created_at=request.POST.get('created_at'),
             updated_at=request.POST.get('update_at')
         )
-        redirect('todolist', id=todolist.pk)
+        return redirect('todolist')
+
+
+def todolist_update(request, pk):
+    todolist = get_object_or_404(ToDoList, id=pk)
+    if request.method == "GET":
+        return render(request, "update.html", {"todolist": todolist})
+    if request.method == 'POST':
+        todolist.description = request.POST.get('description')
+        todolist.detailed_description = request.POST.get('detailed_description')
+        todolist.status = request.POST.get('status')
+        todolist.time = request.POST.get('time')
+        todolist.created_at = request.POST.get('created_at')
+        todolist.updated_at = request.POST.get('update_at')
+
+    return redirect('todolist', pk=todolist.pk)
